+++ ios/Classes/TiPaintPaintViewProxy.m
@@ -5,6 +5,7 @@
  */
 
 #import "TiPaintPaintViewProxy.h"
+#import "TiPaintPaintView.h"
 #import "TiUtils.h"
 
 @implementation TiPaintPaintViewProxy
@@ -17,12 +18,12 @@
 
 -(void)undo:(id)args
 {
-	[[self view] performSelectorOnMainThread:@selector(undo:) withObject:args waitUntilDone:NO];
+	[[self view] performSelectorOnMainThread:@selector(undo) withObject:nil waitUntilDone:NO];
 }
 
 -(void)redo:(id)args
 {
-	[[self view] performSelectorOnMainThread:@selector(redo:) withObject:args waitUntilDone:NO];
+	[[self view] performSelectorOnMainThread:@selector(redo) withObject:nil waitUntilDone:NO];
 }
 
 -(TiBlob*)toBlob:(id)args
@@ -32,6 +33,42 @@
 		result = [[self view] toBlob:args];
 	}, YES);
 	return result;
+}
+
+// Playback methods
+-(void)playbackDrawing:(id)args
+{
+	[[self view] performSelectorOnMainThread:@selector(playbackDrawing:) withObject:args waitUntilDone:NO];
+}
+
+-(void)pausePlayback:(id)args
+{
+	[[self view] performSelectorOnMainThread:@selector(pausePlayback) withObject:nil waitUntilDone:NO];
+}
+
+-(void)resumePlayback:(id)args
+{
+	[[self view] performSelectorOnMainThread:@selector(resumePlayback) withObject:nil waitUntilDone:NO];
+}
+
+-(void)stopPlayback:(id)args
+{
+	[[self view] performSelectorOnMainThread:@selector(stopPlayback) withObject:nil waitUntilDone:NO];
+}
+
+-(void)setPlaybackSpeed:(id)args
+{
+	[[self view] performSelectorOnMainThread:@selector(setPlaybackSpeed:) withObject:args waitUntilDone:NO];
+}
+
+-(NSNumber*)getPlaybackProgress:(id)args
+{
+	__block CGFloat result = 0.0;
+	TiThreadPerformOnMainThread(^{
+		TiPaintPaintView* paintView = (TiPaintPaintView*)[self view];
+		result = [paintView getPlaybackProgress];
+	}, YES);
+	return [NSNumber numberWithFloat:result];
 }
 
 @end